+++ components/sc.js
@@ -1,7 +1,6 @@
 import React from "react";
 import styles from "../styles/SC.module.css"
 import Editor from "../components/editor"
-import CodeEditor from "../components/codeEditor"
 import Arrow from "../components/arrow"
 import lexurgy from "../lib/lexurgy";
 import Checkdrop from "./checkdrop";
@@ -50,13 +49,14 @@ export default class SC extends React.Component {
             styles={`${styles.stackedEditor} ${styles.inputContainer}`}
           />
           <Arrow/>
-          <CodeEditor
+          <Editor
             id="changes"
             label="Sound Changes"
             value={this.state.changes}
             updateValue={this.updateEditorWith}
             expectedFileType=".lsc"
             styles={`${styles.stackedEditor} ${styles.changesContainer}`}
+            showLineNumbers
           />
           <Arrow/>
           <Editor
